+++ preprocessy/pipelines/_feature_selection_pipeline.py
@@ -1,7 +1,7 @@
 from preprocessy.data_splitting import Split
 from preprocessy.encoding import Encoder
 from preprocessy.feature_selection import SelectKBest
-from preprocessy.handlenullvalues import NullValuesHandler
+from preprocessy.missing_data import NullValuesHandler
 from preprocessy.outliers import HandleOutlier
 from preprocessy.parse import Parser
 from preprocessy.scaling import Scaler
